api/requirements.txt
# Celeste Yacht AI - Python Dependencies
# Core FastAPI and web framework dependencies

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP and networking
httpx==0.25.2
requests==2.31.0

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2

# Logging and utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# AI/ML libraries for CLIP-ViT-B/32 and image processing
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0

# PDF and image processing
PyMuPDF==1.23.8
pytesseract==0.3.10

# Optional: Add these later when needed
# openai==1.3.5

# Database (for future data persistence)  
# sqlalchemy==2.0.23
# databases==0.8.0
# psycopg2-binary==2.9.9

# Caching (for future Redis integration)
# redis==5.0.1
# aioredis==2.0.1

# Search engine (for future Elasticsearch integration)
# elasticsearch==8.11.0

# Authentication (for future auth implementation)
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4 